# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AllocationDimension": {
        "description": "Grouping dimension for the current-allocation endpoint.",
        "enum": [
          "entity",
          "currency",
          "kind"
        ],
        "title": "AllocationDimension",
        "type": "string"
      },
      "AllocationResponse": {
        "additionalProperties": false,
        "description": "Current allocation snapshot grouped by one dimension.",
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "by": {
            "$ref": "#/components/schemas/AllocationDimension"
          },
          "slices": {
            "items": {
              "$ref": "#/components/schemas/AllocationSlice"
            },
            "title": "Slices",
            "type": "array"
          }
        },
        "required": [
          "as_of",
          "by",
          "slices"
        ],
        "title": "AllocationResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReturnsScope": {
        "description": "Scope dimension of ``mart_returns_daily`` (ADR-0039).",
        "enum": [
          "account",
          "asset_class",
          "household"
        ],
        "title": "ReturnsScope",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Read-only JSON API over the Penge analytics marts. All amounts are reported in EUR and DKK in parallel; account identifiers are masked server-side. Write-capable surfaces are limited to staged imports under /imports per ADR-0037, Enable Banking consent and sync under /connections per ADR-0040, and the guarded dbt-only refresh under POST /meta/refresh per ADR-0046. The /connections endpoints provide the in-app Enable Banking consent flow and are only active where the EB signing key is configured. POST /meta/refresh triggers the guarded dbt-only shadow-build and atomic-promotion refresh from ADR-0046 without re-syncing any bank connection.",
    "title": "Penge read API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/allocation/current": {
      "get": {
        "description": "Latest-day allocation grouped by entity, currency, or account kind.",
        "operationId": "allocation_current_allocation_current_get",
        "parameters": [
          {
            "in": "query",
            "name": "by",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/AllocationDimension",
              "default": "kind"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Allocation Current"
      }
    },
...
    "/benchmarks/daily": {
      "get": {
        "description": "Daily close series of one benchmark, in its native currency.\n\nAn unknown instrument id yields an empty series, not an error \u2014\nthe UI treats it the same as \"no prices in window\".",
        "operationId": "benchmarks_daily_benchmarks_daily_get",
        "parameters": [
          {
            "description": "Instrument id from /benchmarks.",
            "in": "query",
            "name": "instrument_id",
            "required": true,
            "schema": {
              "description": "Instrument id from /benchmarks.",
              "title": "Instrument Id",
              "type": "string"
            }
          },
          {
            "description": "First day of the window (inclusive). Default: one year ago.",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First day of the window (inclusive). Default: one year ago.",
              "title": "Since"
            }
          },
          {
            "description": "Last day of the window (inclusive). Default: today.",
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last day of the window (inclusive). Default: today.",
              "title": "Until"
            }
          },
          {
            "description": "Page size.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 1000,
              "description": "Page size.",
              "maximum": 10000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Page start offset.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Page start offset.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkSeriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Benchmarks Daily"
      }
    },
    "/cashflow/daily": {
      "get": {
        "description": "Daily cashflow series per account (absent days mean zero).",
        "operationId": "cashflow_daily_cashflow_daily_get",
        "parameters": [
          {
            "description": "First day of the window (inclusive). Default: one year ago.",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First day of the window (inclusive). Default: one year ago.",
              "title": "Since"
            }
          },
          {
            "description": "Last day of the window (inclusive). Default: today.",
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last day of the window (inclusive). Default: today.",
              "title": "Until"
            }
          },
          {
            "description": "Filter to one account id.",
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one account id.",
              "title": "Account Id"
            }
          },
          {
            "description": "Filter to one entity id.",
            "in": "query",
            "name": "entity_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one entity id.",
              "title": "Entity Id"
            }
          },
          {
            "description": "Page size.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 1000,
              "description": "Page size.",
              "maximum": 10000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Page start offset.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Page start offset.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashflowSeriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cashflow Daily"
      }
    },
...
    "/meta/refresh": {
      "post": {
        "description": "Trigger the guarded dbt-only refresh, without re-syncing connections.\n\nReuses the exact shadow-build/test-plus-atomic-promotion path\n(``DbtRunner.refresh``) and the shared advisory lock + durable\npending marker described in ADR-0046, so this route, the manual\nconnection-sync route, import commits, and the scheduled worker can\nnever overlap.\nThe pending marker is created *before* ``refresh()`` runs (mirroring\nthe scheduled worker's own write-intent tracking) so that, if this\nprocess is killed or dbt fails, the next scheduled run still sees\na pending refresh and retries automatically; it is cleared only\nonce promotion succeeds.",
        "operationId": "meta_refresh_meta_refresh_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetaRefreshResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "502": {
            "description": "The shadow dbt build, tests, or schema promotion failed. Live marts are unchanged; the pending marker is preserved (or created, if none existed yet) so the next scheduled run retries."
          },
          "503": {
            "description": "The refresh lock is already held by the scheduled worker, a connection sync, import commit, or another manual trigger, or durable refresh intent could not be persisted."
          }
        },
        "summary": "Meta Refresh"
      }
    },
...
    "/returns/daily": {
      "get": {
        "description": "Daily return factors per scope from ``mart_returns_daily``.",
        "operationId": "returns_daily_returns_daily_get",
        "parameters": [
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ReturnsScope",
              "default": "household"
            }
          },
          {
            "description": "Filter to one scope key (account id, asset-class key, or 'household').",
            "in": "query",
            "name": "scope_key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to one scope key (account id, asset-class key, or 'household').",
              "title": "Scope Key"
            }
          },
          {
            "description": "First day of the window (inclusive). Default: one year ago.",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First day of the window (inclusive). Default: one year ago.",
              "title": "Since"
            }
          },
          {
            "description": "Last day of the window (inclusive). Default: today.",
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last day of the window (inclusive). Default: today.",
              "title": "Until"
            }
          },
          {
            "description": "Page size.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 1000,
              "description": "Page size.",
              "maximum": 10000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Page start offset.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Page start offset.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReturnsSeriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Returns Daily"
      }
    },
...
  }
}